[FEATURE] Profile description and issue templates (#5)
* E-constellatio presentation

Like for branches of E-constellatio, a presentation was written to explain what is E-constellatio and its branches.

* Pull request template

Pull request template for E-cosntellatio was added. It's the same template as Libra fondation with some changes in phrase intonation.

* Issue templates

Issue template was added for bug and new feature request. It's the same issue template as Libra fondation. I don't performed any changes and I don't think any changes is required.

* Spelling correction

* Spelling correction and icons url

* Delete old icons

* Missing icon

* Missing changes

* [FEATURE] Refactor issues templates and add a new one (#4)

* change extension from md to yml

* Update README.yml

* Validations terms update

* indent issue in checkboxes

* Indentation fix part 2

* Fix issue in context part

* Fix issue

* Steps to reproduce issue fix

* Fix indentation issue

* Update checks part

* Delete useless lines

* Delete multiline comment

Multiline description aren't supported in YML

* Delete code of conduct url

* Update name and description of issue template

* Icon issue fix

* Change feature request template extension

* Refactor feature request template

* Reduce description of implementation

The description of implementation was too long to be a valid yml format.

* Create a refactor template

* Delete expected behaviour part

A remove part was missing

* Delete duplicate section

* Fix wrong title in refactor tempalte

* UPPER_CASE title of refactor template

* UPPER_CASE issue template title

* Change refactor label

* Change mandatory part